Move attribute accesses to `TypeSystemInferenceExtensionContext`
`getAttributes`/`hasCustomAttributes`/`getCustomAttributes` from
`TypeSystemContext` are only used when the implementation is also a
`TypeSystemInferenceExtensionContext`.

This commit removes these three functions from `IrTypeSystemContext`
while keeping them unchanged in `ConeInferenceContext` and
`ClassicTypeSystemContext`.

Motivation:
During an analysis of the Kotlin IR tree, there was an unaccountable
and unused instance of `IrElement` found inside `IrTypeSystemContext`.
A minor refactoring could eliminate the need for such a placeholder
and, at the same time, align the definition of `TypeSystemContext` with
its actual use.
